Ensure we use the output of getResolutionsBatch in Compliance
Before this change, we were checking for 'response.ok'. That would have
been OK if we handled the HTTP request ourselves. But since that's
handled by the getResolutionsBatch method, the value returned from that
method is the actual result directly, without any info about the request.

Therefore, 'request.ok' is undefined and makes it appear as if there
were no remediations available, as we are passed {} as the list of
remediations available. To fix this, we simply need to use the method as
intended, and parse the output without checking the status of the request.